pyproject.toml
[project]
name = "langgraph-cot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0", # Change 1.1.6 to 0.2.0
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "openai>=2.14.0",
]
